+++ cookbooks/browsers/recipes/chromium.rb
@@ -26,16 +26,3 @@ install_app "chromium" do
   profile browsers['profiles']['chromium']
 end
 
-# Unity Web Apps install
-if platform?("ubuntu") and platform_version == 12.04
-  codename = platform_codename
-
-  core_ppa "webapps-preview" do
-    uri "ppa:webapps/preview"
-    distribution codename
-    action :add
-  end
-
-  package "unity-webapps-preview"
-end
-